requirements.txt
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.2.0
oracledb==2.1.1
opencv-python==4.9.0.80
spacy==3.7.4
watchdog==4.0.0
regex==2023.12.25
ultralytics 
opencv-python
pdfminer.six
PyMuPDF
pdf2image
pandas
rapidfuzz

